시간표가 하나뿐인 경우 삭제 불가능하게 함

+++ app/src/main/java/com/wafflestudio/snutt2/views/logged_in/home/drawer/TableMoreActionBottomSheet.kt
@@ -58,9 +58,11 @@ fun TableMoreActionBottomSheet(
             text = stringResource(R.string.home_drawer_table_delete),
         ) {
             scope.launch {
-                showTableDeleteDialog(table.id, composableStates) { tableId ->
-                    tableListViewModel.deleteTable(tableId)
-                }
+                if (tableListViewModel.checkTableDeletable()) {
+                    showTableDeleteDialog(table.id, composableStates) { tableId ->
+                        tableListViewModel.deleteTable(tableId)
+                    }
+                } else context.toast(context.getString(R.string.home_drawer_delete_table_unable_alert_message))
             }
         }
         MoreActionItem(
